move EventAlerts to modals folder
