+++ frontend/src/modules/modals/index.ts
@@ -4,6 +4,7 @@
  * @file Components which can be reused across app
  *
  */
+export * from './EventAlerts';
 export * from './MultiStepWizard';
 export * from './ModalPopup';
 export * from './DiscardAlarmLimitsContent';
